refactor: forms folder
-Reorganize ui folder for improved readability/organization.
